Fix all clippy warnings for CI compliance
- Use iterators instead of index loops (needless_range_loop)
- Replace manual abs_diff with .abs_diff() method
- Replace manual range checks with .contains() pattern
- Replace manual clamp with .clamp() method
- Use .is_multiple_of() instead of % 2 == 0
- Add #[derive(Default)] where appropriate
- Collapse nested if statements
- Remove unnecessary casts and returns
- Add #[allow] for intentional patterns (too_many_arguments, arc_with_non_send_sync)

+++ src/game_repr/piece_moves/pawn.rs
@@ -56,8 +56,8 @@ impl Position {
         // En passant - only check if there was a previous move
         if !self.prev_moves.is_empty() {
             let prev_move = self.prev_moves.last().unwrap();
-            let prev_to = prev_move._to() as usize;
-            let prev_from = prev_move._from() as usize;
+            let prev_to = prev_move._to();
+            let prev_from = prev_move._from();
 
             // Check if last move was a two-square pawn advance
             if self.position[prev_to].piece_type == Type::Pawn
